+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.2 (WIP)
 
 * Fix Client class methods not being set on the server-side for classes defined in a plugin
+* Add `Criteria#isEmpty()` field check
 
 ## 1.1.1 (2020-07-24)